Server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.9.0
sqlalchemy==2.0.35
pg8000==1.31.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
httpx==0.27.0
PyPDF2==3.0.1
python-docx==1.1.2
pdfplumber==0.11.4
